+++ include/utils/SkyboxShader.h
@@ -26,7 +26,7 @@ private:
 
 	std::string LoadShader(const std::string& fileName);
 	void CheckShaderError(GLuint shader, GLuint flag, bool isProgram, const std::string& errorMessage);
-	GLuint CreateShader(const std::string& text, unsigned int type);
+	GLuint CreateShader(const std::string& text, size_t type);
 
 	GLuint m_program;
 	std::vector< GLuint > m_shaderids;
