+++ include/utils/Util.h
@@ -11,9 +11,9 @@ namespace Util
         std::vector<std::string> elems;
         
         const char* cstr = s.c_str();
-        unsigned int strLength = s.length();
-        unsigned int start = 0;
-        unsigned int end = 0;
+        size_t strLength = s.length();
+        size_t start = 0;
+        size_t end = 0;
         
         while(end <= strLength)
         {
